src/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.4
requests==2.31.0
google-generativeai>=0.8.0
python-dotenv==1.0.0
pydantic==2.5.0
motor==3.3.2
pymongo==4.6.0